all checkpoints done without user checking for card/user methods
