+++ controllers/auth.js
@@ -1,0 +1,36 @@
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const User = require('../models/user');
+
+module.exports.createUser = (req, res) => {
+  const {
+    name, about, avatar, email,
+  } = req.body;
+
+  bcrypt.hash(req.body.password, 10)
+    .then((hash) => User.create({
+      name, about, avatar, email, password: hash,
+    }))
+    .then(() => res.status(201).json({
+      name, about, avatar, email,
+    }))
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        res.status(400).json({ message: err.message });
+      } else {
+        res.status(500).json({ message: 'Server error' });
+      }
+    });
+};
+
+module.exports.login = (req, res) => {
+  const { email } = req.body;
+  return User.findOne({ email }).select('+password')
+    .then((user) => {
+      const token = jwt.sign({ _id: user._id }, 'some-secret-key', { expiresIn: '7d' });
+      res.status(200).send({ token });
+    })
+    .catch((err) => {
+      res.status(401).send({ message: err.message });
+    });
+};
