+++ package.json
@@ -21,6 +21,7 @@
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.9.24",
+    "mongoose-unique-validator": "^2.0.3",
     "validator": "^13.1.1"
   }
 }
